+++ src/routes/graphql/types/inputTypes.ts
@@ -4,7 +4,6 @@ import {
   GraphQLInputObjectType,
   GraphQLInt,
   GraphQLNonNull,
-  GraphQLObjectType,
   GraphQLString,
 } from 'graphql';
 import { UUIDType } from './uuid.js';
